working on walkready when penalized
